+++ app/(main)/listogs/listogs.module.css
@@ -159,6 +159,21 @@
     background-color: darkred;
 }
 
+.button {
+    margin-top: 10px;
+    background-color: #007bff; /* Blue background */
+    color: white; /* White text */
+    border: none;
+    border-radius: 4px; /* Rounded corners */
+    padding: 10px 20px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+.button:hover {
+    background-color: #0056b3; /* Darker blue on hover */
+}
+
 .progressContainer {
     margin-top: 10px;
     padding: 10px;
@@ -203,89 +218,55 @@
     flex-wrap: wrap;
     gap: 20px;
     margin-top: 20px;
+    justify-content: flex-start;
 }
 
 .youtubeEmbed {
-    flex: 1 1 calc(50% - 20px); /* Adjust width for two columns */
-    max-width: 560px; /* Ensure it doesn't exceed iframe width */
-}
-
-.youtubeLink {
-    display: block;
+    flex: 1 1 400px;
+    max-width: 420px;
+    min-width: 280px;
+    box-sizing: border-box;
     margin-bottom: 10px;
-    color: #007bff;
-    text-decoration: none;
-    word-wrap: break-word;
-    padding: 8px;
-    background-color: #f8f9fa;
-    border-radius: 4px;
-    transition: background-color 0.2s;
+    background: #fff;
+    border-radius: 8px;
+    box-shadow: 0 1px 4px rgba(0,0,0,0.04);
+    padding: 12px 12px 0 12px;
+    display: flex;
+    flex-direction: column;
+    align-items: stretch;
 }
 
-.youtubeLink:hover {
-    background-color: #e9ecef;
+.youtubeDivider {
+    border: none;
+    border-bottom: 1.5px solid #e0e0e0;
+    margin: 18px 0 10px 0;
+    width: 100%;
+}
+
+.discogsTitleLink {
+    color: #007bff;
+    text-decoration: underline;
+    font-weight: bold;
+    font-size: 1.05em;
+    transition: color 0.2s;
+    cursor: pointer;
+}
+
+.discogsTitleLink:hover {
+    color: #0056b3;
     text-decoration: underline;
 }
 
-.videoResponsive {
-    overflow: hidden;
-    padding-bottom: 56.25%; /* 16:9 aspect ratio */
-    position: relative;
-    height: 0;
-}
-
-.videoResponsive iframe {
-    left: 0;
-    top: 0;
-    height: 100%;
-    width: 100%;
-    position: absolute;
-}
-
-.devSliderContainer {
-    margin-top: 10px;
-    display: flex;
-    align-items: center;
-}
-
-.devSliderLabel {
-    font-size: 16px;
-    color: #333;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.devSlider {
-    width: 40px;
-    height: 20px;
-    appearance: none;
-    background-color: #ccc;
-    border-radius: 10px;
-    position: relative;
-    outline: none;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.devSlider:checked {
-    background-color: #007bff;
-}
-
-.devSlider::before {
-    content: '';
-    position: absolute;
-    width: 18px;
-    height: 18px;
-    background-color: white;
-    border-radius: 50%;
-    top: 1px;
-    left: 1px;
-    transition: transform 0.3s;
-}
-
-.devSlider:checked::before {
-    transform: translateX(20px);
+@media (max-width: 900px) {
+    .youtubeContainer {
+        flex-direction: column;
+        gap: 10px;
+    }
+    .youtubeEmbed {
+        max-width: 100%;
+        min-width: 0;
+        width: 100%;
+    }
 }
 
 .taskList {
@@ -515,4 +496,58 @@
     font-size: 14px;
     color: #666;
     margin: 10px 0;
+}
+
+.socketStatus {
+    margin-bottom: 10px;
+    font-size: 14px;
+}
+
+.connected {
+    color: #28a745;
+    font-weight: bold;
+}
+
+.disconnected {
+    color: #dc3545;
+    font-weight: bold;
+}
+
+.artistName {
+    font-size: 18px;
+    font-weight: bold;
+    margin-bottom: 10px;
+    color: #007bff;
+}
+
+.rateLimitedTask {
+    border-color: #fd7e14;
+    background-color: rgba(253, 126, 20, 0.05);
+}
+
+.rateLimitedTask.selectedTask {
+    border-color: #fd7e14;
+    box-shadow: 0 0 4px rgba(253, 126, 20, 0.3);
+    background-color: rgba(253, 126, 20, 0.15);
+}
+
+.playlistInfo {
+    font-size: 14px;
+    color: #666;
+    margin: 10px 0;
+}
+
+.socketStatus {
+    margin-bottom: 10px;
+    font-size: 14px;
+}
+
+.connected {
+    color: #28a745;
+    font-weight: bold;
+}
+
+.disconnected {
+    color: #dc3545;
+    font-weight: bold;
 }
