+++ package.json
@@ -27,7 +27,9 @@
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^19.0.0",
     "readline": "^1.3.0",
-    "sharp": "^0.34.1"
+    "sharp": "^0.34.1",
+    "socket.io": "^4.8.1",
+    "socket.io-client": "^4.8.1"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
